Introduce IExpressionConverterFactory for extensibility
Very coarse extensibility of the library's capabilities to convert LINQ
expression trees into RethinkDB terms.  However, this first step will
also allow for the storage of state and configuration on an implementation
of IExpressionConverterFactory, which will be necessary for more
fine-grained extensibility to be added without the use of global static
state for registration of it.