Use IGroupingDictionary for $reql_type$=GROUPED_DATA
IGroupingDictionary is a type that's only supported for read, it's
automatically returned from Group() queries and grouped aggregations, and
it's also used by rethinkdb-net-newtonsoft because the GROUPED_DATA
reql_type has a special format that can't be intrinsically guessed by the
JSON parser.  IGroupingDictionary<,> is an IDictionary<,> so it shouldn't
even be noticed by the consumers.