Stash totally unpleasant work on using the JSON protocol
The state of JSON in .NET is a bit abysmal.  I don't want to add new
third-party dependencies like JSON.NET, the .NET's own JSON support
is highly coupled to using data contracts rather than free-form objects,
and I can't find a decent, simple parser that I can incorporate with
the correct licensing and an approach that doesn't build an object model
which I would have to translate into a Response object model.

On the plus side, I benchmarked a simple JsonWriter against the Protobuf
library for writing queries.  It showed a bit of a speed-up, but not
enough to salivate over.  I'm shelving this work until RethinkDB threatens
to kill the protobuf protocol. :-)