Import JSON library from sinesignal/ottoman @dragan for JSON protocol
This branch is now partially working with the JSON protocol; some tests
pass.  There's obviously a few mistakes in that all tests don't pass,
and there are a few FIXMEs for things that I've half-assed to get a start
on, and there's a lot of code added to Connection that probably doesn't
belong there.

It might be a good idea to separate out the protocol details from the
Connection class entirely, and make rethinkdb-net capable of using
either protocol.  Mostly that will make it clear what code is related
to the JSON protocol, but it would also be easy to support both in
the driver if there was a pressing need to.