Work-in-progress on DateTimeOffset constructors
A complicated constructor to support new DateTimeOffset(dt, offset) that
converts TimeSpans to offsets, server-side, by formatting the correct
offset string.

It seems like the DateTimeOffsetDatumConverter isn't handling the offsets
the way I would expect; I've made a change here, but it breaks the tests.
The tests need to be re-evaluated to ensure my change is accurate...
it might not be.