Cleanup duplicate MemberAccess code
Now that BaseExpression supports MemberAccess on any type, the
single/two parameter lambda converters don’t have to have duplicate
code doing the same thing.  They still need a special case to handle
Convert nodes on the parameter object.

GenericTypeConstraintExpressionTests required a change; accessing a
member on the interface here doesn’t seem right because we don’t know
how to treat that interface.  So, I supplied the correct datum
converter factory for that interface.  Code makes more sense now, and
tests pass.  For the original reported issue #169 it should be
functionally equivalent as that issue was using the newtonsoft
converter which presumably would work on the interface as well.